test(workspace): split workspace.test.ts to satisfy max-lines cap
Extract the loadWorkspaceBootstrapFiles and readWorkspaceFileWithGuards
describe suites into workspace.bootstrap-files.test.ts so each test file
stays under the 1000 code-line lint cap. No test logic changed.
